Flytt opprett oppgave-funksjonalitet til vilkårsvurdering, der det hører hjemme. Fjern overflødig skalOppretteOppgave-funksjon
