+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/steg/Vilkårsvurdering.kt
@@ -10,16 +10,23 @@ import no.nav.familie.ba.sak.behandling.grunnlag.personopplysninger.Persongrunnl
 import no.nav.familie.ba.sak.behandling.grunnlag.søknad.SøknadGrunnlagService
 import no.nav.familie.ba.sak.behandling.vedtak.VedtakService
 import no.nav.familie.ba.sak.behandling.vilkår.BehandlingResultatService
+import no.nav.familie.ba.sak.behandling.vilkår.BehandlingResultatType
 import no.nav.familie.ba.sak.behandling.vilkår.SakType.Companion.hentSakType
 import no.nav.familie.ba.sak.behandling.vilkår.VilkårService
 import no.nav.familie.ba.sak.beregning.BeregningService
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.RessursUtils
 import no.nav.familie.ba.sak.common.toPeriode
-import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
+import no.nav.familie.ba.sak.config.FeatureToggleService
+import no.nav.familie.ba.sak.task.OpprettOppgaveTask
+import no.nav.familie.kontrakter.felles.oppgave.Oppgavetype
+import no.nav.familie.prosessering.domene.TaskRepository
 import no.nav.nare.core.evaluations.Resultat
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import java.time.LocalDate
 
 @Service
 class Vilkårsvurdering(
@@ -29,17 +36,19 @@ class Vilkårsvurdering(
         private val beregningService: BeregningService,
         private val persongrunnlagService: PersongrunnlagService,
         private val behandlingResultatService: BehandlingResultatService,
-        private val behandlingService: BehandlingService
+        private val behandlingService: BehandlingService,
+        private val featureToggleService: FeatureToggleService,
+        private val taskRepository: TaskRepository
 ) : BehandlingSteg<String> {
 
     @Transactional
     override fun utførStegOgAngiNeste(behandling: Behandling,
                                       data: String): StegType {
         val personopplysningGrunnlag = persongrunnlagService.hentAktiv(behandling.id)
-                                       ?: error("Fant ikke personopplysninggrunnlag på behandling ${behandling.id}")
+                ?: error("Fant ikke personopplysninggrunnlag på behandling ${behandling.id}")
 
         val behandlingResultat = behandlingResultatService.hentAktivForBehandling(behandlingId = behandling.id)
-                                 ?: throw Feil("Fant ikke aktiv behandlingresultat på behandling ${behandling.id}")
+                ?: throw Feil("Fant ikke aktiv behandlingresultat på behandling ${behandling.id}")
 
         if (behandling.opprinnelse == BehandlingOpprinnelse.AUTOMATISK_VED_FØDSELSHENDELSE) {
             vilkårService.vurderVilkårForFødselshendelse(behandling.id)
@@ -55,6 +64,22 @@ class Vilkårsvurdering(
 
         if (behandling.opprinnelse == BehandlingOpprinnelse.AUTOMATISK_VED_FØDSELSHENDELSE) {
             behandlingService.oppdaterStatusPåBehandling(behandling.id, BehandlingStatus.GODKJENT)
+
+            val aktivtBehandlingResultat = behandlingResultatService.hentAktivForBehandling(behandlingId = behandling.id)
+                    ?: throw Feil("Fant ikke aktiv behandlingresultat på behandling ${behandling.id}")
+
+            if (aktivtBehandlingResultat.hentSamletResultat() != BehandlingResultatType.INNVILGET
+                    && featureToggleService.isEnabled("familie-ba-sak.lag-oppgave")
+                    && !featureToggleService.isEnabled("familie-ba-sak.rollback-automatisk-regelkjoring")) {
+                val nyTask = OpprettOppgaveTask.opprettTask(
+                        behandlingId = behandling.id,
+                        oppgavetype = Oppgavetype.BehandleSak,
+                        fristForFerdigstillelse = LocalDate.now()
+                )
+                taskRepository.save(nyTask)
+            } else {
+                LOG.info("Lag opprettOppgaveTask er skrudd av i miljø eller behandlingen av fødselshendelsen var innvilget")
+            }
         }
 
         return hentNesteStegForNormalFlyt(behandling)
@@ -67,7 +92,7 @@ class Vilkårsvurdering(
     override fun postValiderSteg(behandling: Behandling) {
         if (behandling.type != BehandlingType.TEKNISK_OPPHØR && behandling.type != BehandlingType.MIGRERING_FRA_INFOTRYGD_OPPHØRT) {
             val behandlingResultat = vilkårService.hentVilkårsvurdering(behandlingId = behandling.id)
-                                     ?: error("Finner ikke vilkårsvurdering på behandling ved validering.")
+                    ?: error("Finner ikke vilkårsvurdering på behandling ved validering.")
 
             val listeAvFeil = mutableListOf<String>()
 
@@ -79,9 +104,9 @@ class Vilkårsvurdering(
 
             val harGyldigePerioder = periodeResultater.any { periodeResultat ->
                 periodeResultat.allePåkrevdeVilkårVurdert(PersonType.SØKER,
-                                                          sakType) &&
-                periodeResultat.allePåkrevdeVilkårVurdert(PersonType.BARN,
-                                                          sakType)
+                        sakType) &&
+                        periodeResultat.allePåkrevdeVilkårVurdert(PersonType.BARN,
+                                sakType)
             }
 
             when {
@@ -103,7 +128,7 @@ class Vilkårsvurdering(
                                 listeAvFeil.add("Vilkår '${vilkårResultat.vilkårType}' for barn med fødselsdato ${barn.fødselsdato} har fra-og-med dato før barnets fødselsdato.")
                             }
                             if (vilkårResultat.periodeFom != null &&
-                                vilkårResultat.toPeriode().fom.isAfter(barn.fødselsdato.plusYears(18))) {
+                                    vilkårResultat.toPeriode().fom.isAfter(barn.fødselsdato.plusYears(18))) {
                                 listeAvFeil.add("Vilkår '${vilkårResultat.vilkårType}' for barn med fødselsdato ${barn.fødselsdato} har fra-og-med dato etter barnet har fylt 18.")
                             }
                         }
@@ -111,10 +136,14 @@ class Vilkårsvurdering(
 
             if (listeAvFeil.isNotEmpty()) {
                 throw Feil(message = "Validering av vilkårsvurdering feilet for behandling ${behandling.id}",
-                           frontendFeilmelding = RessursUtils.lagFrontendMelding("Vilkårsvurderingen er ugyldig med følgende feil:",
-                                                                                 listeAvFeil)
+                        frontendFeilmelding = RessursUtils.lagFrontendMelding("Vilkårsvurderingen er ugyldig med følgende feil:",
+                                listeAvFeil)
                 )
             }
         }
+    }
+
+    companion object {
+        val LOG: Logger = LoggerFactory.getLogger(Vilkårsvurdering::class.java)
     }
 }
